add a sortArrayByBulkId utility file.
